Point & Position updated
Point made to be just a collidable extension of Position.  Otherwise, functionality is the same.  Path class updated to match changes in Point. Positon fixed operator overloads, move semantics, and conversions for and eliminated the magnitude and direction so that Position is just a 4 float value.  This allows for matrix operations.